rlottie: Fix cmake build

+++ src/vector/vline.h
@@ -55,7 +55,7 @@ inline float VLine::angle() const
     const float dx = mX2 - mX1;
     const float dy = mY2 - mY1;
 
-    const float theta = std::atan2f(dy, dx) * 180.0f / K_PI;
+    const float theta = std::atan2(dy, dx) * 180.0f / K_PI;
     return theta;
 }
